requirements.txt
seleniumbase==4.7.0
pytest
python-dotenv==1.0.0
PyYAML==6.0
requests
selenium~=4.5.0
webdriver-manager==3.8.6
pytest-fixtures==0.1.0
pytest-html>=2.0.1,<3.0.0
pytest-metadata
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest-reportlog

py~=1.11.0
colorama~=0.4.6
beautifulsoup4~=4.11.1
soupsieve~=2.3.2.post1
h11~=0.14.0
cffi~=1.15.1
Pillow~=10.0.0
ipython~=7.34.0
pip~=23.1.2
attrs~=23.1.0
wheel~=0.40.0
cryptography~=38.0.1
ipdb~=0.13.9
decorator~=5.1.1
parso~=0.8.3
jedi~=0.18.1
nose~=1.3.7
setuptools~=67.8.0
rich~=12.6.0
Pygments~=2.13.0
commonmark~=0.9.1
toml~=0.10.2
trio~=0.22.0
outcome~=1.2.0
sniffio~=1.3.0
exceptiongroup
sortedcontainers~=2.4.0
idna~=3.4
behave~=1.2.6
execnet~=1.9.0
six~=1.16.0
parse~=1.19.0
wcwidth~=0.2.6
pluggy~=1.0.0
iniconfig~=2.0.0
pexpect~=4.8.0
certifi~=2022.12.7
chardet~=5.0.0
traitlets~=5.9.0
packaging~=23.1
backcall~=0.2.0
pickleshare~=0.7.5
ptyprocess~=0.7.0
urllib3~=1.26.12
cssselect~=1.2.0
fasteners~=0.18
pycparser~=2.21
pyparsing~=3.0.9
zipp~=3.15.0
platformdirs~=3.5.3
parameterized~=0.8.1
wsproto~=1.2.0
sbvirtualdisplay~=1.1.0
tqdm~=4.65.0